requirements.txt
Django
djangorestframework
django-cors-headers
PyJWT[crypto]
psycopg[binary]
requests
django-environ
cryptography
drf-spectacular
python-dateutil
confluent-kafka
celery
redis
django-redis
black
isort
flake8
mypy
django-stubs
pytest
pytest-django
google-auth-oauthlib
oauthlib
requests-oauthlib
types-requests
types-python-dateutil
pytest-json-report
google-genai
grpcio
grpcio-tools
grpcio-status
protobuf
babel
fastembed
numpy==1.26.4
qdrant-client==1.9.1
django-prometheus
prometheus-client
sentry-sdk[django]
gunicorn==23.0.0
gevent>=24.2.1
greenlet>=3.0.0
tenacity>=8.2.0
